feat: simplify mobile menu and move search to header
- Remove BarefootJS title from bottom sheet
- Remove search button from bottom sheet
- Add close button next to drag handle
- Add search icon to mobile header (opens command palette)
- Adjust navigation height for more content visibility
